Upgrade & de-vendor versioneer
